Base form code for create a user
Create a "create a user" form page that will take in forms and display them when the page is called. Create a lists file that will hold lists that are repeated throughout the code. Currently it houses major list and gender list. Also create a form file that will hold all of the from classes