Updated End Semester
This update allows for manually changing the semester via a Forgien
Key field in Usage model that references a Semester Model. The
Semester model currently stores year and season and a bool for
if it is current. Also updated machine usage to default to the
current semester. There should be only one semester with current
set to true